refactor: decouple general interface.ts aggregations and standalone patch files into specific kebab-case boundaries
