+++ src/client/mod.ts
@@ -1,4 +1,4 @@
-export * from "./interface.ts";
+export * from "./client-interface.ts";
 export * from "./client.ts";
 export * from "./quad-store/mod.ts";
 export * from "./sparql-engine/mod.ts";
