+++ src/client/quad-store/mod.ts
@@ -1,4 +1,4 @@
-export * from "./interface.ts";
+export * from "./quad-store-interface.ts";
 export * from "./rdfjs-quad-store.ts";
-export * from "./patch.ts";
+export * from "./patch-queue.ts";
 export * from "./patch-queue-interface.ts";
