+++ src/client/quad-store/quad-store-interface.ts
