+++ src/client/search-index/embedding-service/embedding-service.ts
