+++ src/client/search-index/embedding-service/mod.ts
@@ -1,2 +1,2 @@
-export * from "./interface.ts";
+export * from "./embedding-service.ts";
 export * from "./fake-embedding-service.ts";
