+++ src/client/search-index/mod.ts
@@ -1,1 +1,1 @@
-export * from "./interface.ts";
+export * from "./search-index-interface.ts";
