+++ src/client/search-index/search-index-interface.ts
