+++ src/client/sparql-engine/mod.ts
@@ -1,1 +1,1 @@
-export * from "./interface.ts";
+export * from "./sparql-engine-interface.ts";
