+++ src/client/sparql-engine/sparql-engine-interface.ts
